vetdoc views added, servicios refactored
