+++ src/components/vetdoc/index.docvetcomp.tsx
@@ -1,21 +1,15 @@
+export * from './view/ServiciosActivos';
+export * from './view/ServiciosSeleccion';
+export * from './view/Sidebar';
+
 export * from './CompletarServicioModal';
-
 export * from './ConsultaForm';
-
 export * from './PeluqueriaForm';
-
 export * from './ReservacionesPendientes';
-
 export * from './ServiciosActivos';
-
 export * from './ServiciosCompletados';
-
 export * from './VacunaForm';
-
 export * from './VacunasList';
-
 export * from './VacunationForm';
-
 export * from './VacunationList';
-
 export * from './VetHeader';
