+++ src/pages/vetdoc/servicios.tsx
@@ -4,14 +4,11 @@ import { useRouter } from 'next/router';
 import { logout } from '@/utils/index.utils';
 import { useAuth } from '@/hooks/index.hooks';
 import { Button } from '@/components/ui/index.ui';
-import type { ReservacionV } from '@/types/vetdoc';
-import { Stethoscope, LogOut, Plus, Clipboard, Search, Activity, Scissors, 
-    BedDouble, TestTube, Syringe, History, FileCheck, CheckCircle2} from 'lucide-react';
-import { ReservacionesPendientes, ConsultaForm, PeluqueriaForm, ServiciosActivos, ServiciosCompletados } from '@/components/vetdoc/index.docvetcomp';
+import { Stethoscope, LogOut, Search } from 'lucide-react';
+import type { MainView, ReservacionV, ServiceType } from '@/types/vetdoc';
+import { ReservacionesPendientes, ConsultaForm, PeluqueriaForm, ServiciosCompletados, 
+    ServicioSelection, Sidebar, ServiciosActivosView} from '@/components/vetdoc/index.docvetcomp';
 
-
-type MainView = 'nuevo' | 'activos' | 'historial' | 'completados';
-type ServiceType = 'consulta' | 'peluqueria' | 'internacion' | 'analisis' | 'cirugia';
 
 const ServiciosPage: React.FC = () => {
     const router = useRouter();
@@ -26,7 +23,6 @@ const ServiciosPage: React.FC = () => {
     if (!isAuthenticated) {
         return <div className="flex justify-center items-center h-screen">Acceso Denegado</div>;
     }
-
     const handleLogout = () => {
         logout(router);
     };
@@ -73,9 +69,7 @@ const ServiciosPage: React.FC = () => {
                                         setSelectedReservacion(null);
                                         setSelectedService(null);
                                     }}
-                                    onCancel={() => {
-                                        setSelectedReservacion(null);
-                                    }} 
+                                    onCancel={ () => { setSelectedReservacion(null); }} 
                                 />
                             </div>
                         );
@@ -124,76 +118,15 @@ const ServiciosPage: React.FC = () => {
                                         setSelectedReservacion(null);
                                         setSelectedService(null);
                                     }}
-                                    onCancel={() => {
-                                        setSelectedReservacion(null);
-                                    }} 
+                                    onCancel={ () => {setSelectedReservacion(null); }} 
                                 />
                             </div>
                         );
                     }
                 }
-                return (
-                    <div className="p-6">
-                        <h2 className="text-2xl font-bold mb-6">Nuevo Servicio</h2>
-                        {selectedService ? (
-                            // Aquí irá el formulario específico del servicio seleccionado
-                            <div className="bg-white rounded-lg shadow-md p-6">
-                                <h3 className="text-xl font-semibold mb-4">Formulario de {selectedService}</h3>
-                                {/* Formulario específico se renderizará aquí */}
-                            </div>
-                        ) : (
-                            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-                                <ServiceCard
-                                    icon={<Activity size={40} />}
-                                    title="Consulta Médica"
-                                    description="Registro de consulta veterinaria"
-                                    onClick={() => setSelectedService('consulta')}
-                                />
-                                <ServiceCard
-                                    icon={<Scissors size={40} />}
-                                    title="Peluquería"
-                                    description="Registro de servicio de peluquería"
-                                    onClick={() => setSelectedService('peluqueria')}
-                                />
-                                <ServiceCard
-                                    icon={<BedDouble size={40} />}
-                                    title="Internación"
-                                    description="Registro de internación"
-                                    onClick={() => setSelectedService('internacion')}
-                                />
-                                <ServiceCard
-                                    icon={<TestTube size={40} />}
-                                    title="Análisis Clínico"
-                                    description="Registro de análisis clínico"
-                                    onClick={() => setSelectedService('analisis')}
-                                />
-                                <ServiceCard
-                                    icon={<Syringe size={40} />}
-                                    title="Cirugía"
-                                    description="Registro de cirugía"
-                                    onClick={() => setSelectedService('cirugia')}
-                                />
-                            </div>
-                        )}
-                    </div>
-                );
+                return ( <ServicioSelection onServiceSelect={setSelectedService} /> );
             case 'activos':
-                return (
-                    <div className="p-6">
-                        <div className="mb-6">
-                            <h2 className="text-2xl font-bold mb-4">Servicios Activos</h2>
-                            <div className="flex gap-4 mb-6">
-                                <Button 
-                                    variant="outline" 
-                                    onClick={() => setMainView('nuevo')}
-                                >
-                                    Nuevo Servicio
-                                </Button>
-                            </div>
-                        </div>
-                        <ServiciosActivos />
-                    </div>
-                );
+                return <ServiciosActivosView onNewService={() => setMainView('nuevo')} />;
             case 'completados':
                 return (
                     <div className="p-6">
@@ -247,68 +180,14 @@ const ServiciosPage: React.FC = () => {
                     </Button>
                 </div>
             </header>
-
             {/* Main Content */}
             <div className="flex h-[calc(100vh-4rem)]">
-                {/* Sidebar */}
-                <aside className="w-75 bg-white shadow-lg border-r border-gray-400">
-                    <nav className="p-4">
-                        <div className="space-y-2">
-                            <NavButton
-                                icon={<Plus size={20} />}
-                                text="Nuevo Servicio"
-                                active={mainView === 'nuevo'}
-                                onClick={() => {
-                                    setMainView('nuevo');
-                                    setSelectedService(null);
-                                }}
-                            />
-                            <NavButton
-                                icon={<Clipboard size={20} />}
-                                text="Servicios Activos"
-                                active={mainView === 'activos'}
-                                onClick={() => {
-                                    setMainView('activos');
-                                    setSelectedService(null);
-                                }}
-                            />
-                            <NavButton
-                                icon={<CheckCircle2 size={20} />}
-                                text="Servicios Completados"
-                                active={mainView === 'completados'}
-                                onClick={() => setMainView('completados')}
-                            />
-                            <NavButton
-                                icon={<History size={20} />}
-                                text="Historial Clínico"
-                                active={mainView === 'historial'}
-                                onClick={() => {
-                                    setMainView('historial');
-                                    setSelectedService(null);
-                                }}
-                            />
-                        </div>
-
-                        {/* Subnavegación contextual */}
-                        {mainView === 'nuevo' && selectedService && (
-                            <div className="mt-8 pt-4 border-t">
-                                <h3 className="text-sm font-medium text-gray-500 mb-2">
-                                    Servicio Seleccionado
-                                </h3>
-                                <Button 
-                                    variant="ghost" 
-                                    className="w-full justify-start text-blue-600"
-                                    onClick={() => setSelectedService(null)}
-                                >
-                                    <FileCheck size={16} className="mr-2" />
-                                    Volver a Servicios
-                                </Button>
-                            </div>
-                        )}
-                    </nav>
-                </aside>
-
-                {/* Área de contenido */}
+                <Sidebar
+                    mainView={mainView}
+                    selectedService={selectedService}
+                    onViewChange={setMainView}
+                    onServiceClear={() => setSelectedService(null)}
+                />
                 <main className="flex-1 overflow-auto bg-gray-50">
                     {renderMainContent()}
                 </main>
@@ -316,46 +195,5 @@ const ServiciosPage: React.FC = () => {
         </div>
     );
 };
-
-interface NavButtonProps {
-    icon: React.ReactNode;
-    text: string;
-    active: boolean;
-    onClick: () => void;
-}
-
-const NavButton: React.FC<NavButtonProps> = ({ icon, text, active, onClick }) => (
-    <button
-        onClick={onClick}
-        className={`
-            w-full flex items-center px-4 py-2 rounded-lg
-            transition-colors duration-150 ease-in-out
-            ${active 
-                ? 'bg-blue-50 text-blue-600' 
-                : 'text-gray-600 hover:bg-gray-50'}
-        `}
-    >
-        {icon}
-        <span className="ml-3">{text}</span>
-    </button>
-);
-
-interface ServiceCardProps {
-    icon: React.ReactNode;
-    title: string;
-    description: string;
-    onClick: () => void;
-}
-
-const ServiceCard: React.FC<ServiceCardProps> = ({ icon, title, description, onClick }) => (
-    <div 
-        onClick={onClick}
-        className="bg-white p-6 rounded-lg shadow-md transition-transform hover:scale-105 cursor-pointer"
-    >
-        <div className="text-blue-500 mb-4">{icon}</div>
-        <h3 className="text-xl font-semibold mb-2">{title}</h3>
-        <p className="text-gray-600">{description}</p>
-    </div>
-);
 
 export default ServiciosPage;
